+++ src/lib/Libattr/attr_fn_acl.c
@@ -495,7 +495,7 @@ set_allacl(struct attribute *attr, struct attribute *new, enum batch_op op, int 
 						nsize = strlen(pas->as_string[i]) + 1;
 						pc = pas->as_string[i] + nsize;
 						need = pas->as_next - pc;
-						(void)memmove(pas->as_string[i], pc, (int)need);
+						(void)memmove(pas->as_string[i], pc, (size_t)need);
 						pas->as_next -= nsize;
 						/* compact pointers */
 						for (++i; i < pas->as_npointers; i++)
